refactor(preset): move the preset logic to ProjectsRepository interface
